+++ src/libdm/dm_plugins.cpp
@@ -319,8 +319,8 @@ fb_put_platform_specific(struct fb_platform_specific *fb_p)
 
 
 #define Malloc_Bomb(_bytes_)                                    \
-    fb_log("\"%s\"(%d) : allocation of %lu bytes failed.\n",    \
-           __FILE__, __LINE__, _bytes_)
+    fb_log("\"%s\"(%d) : allocation of %zu bytes failed.\n",    \
+           __FILE__, __LINE__, (size_t)_bytes_)
 
 /**
  * True if the non-null string s is all digits
