+++ src/liboptical/sh_light.c
@@ -174,8 +174,8 @@ light_cvt_visible(const struct bu_structparse *sdp,
     if (optical_debug & OPTICAL_DEBUG_LIGHT) {
 	bu_log("light_cvt_visible(%s, %zu)\n", name, sdp->sp_offset);
 	bu_log("visible: %lu invisible: %lu\n",
-	       LIGHT_O(lt_visible),
-	       LIGHT_O(lt_invisible));
+	       (long unsigned)LIGHT_O(lt_visible),
+	       (long unsigned)LIGHT_O(lt_invisible));
     }
     if (sdp->sp_offset == LIGHT_O(lt_invisible)) {
 	lsp->lt_visible = !lsp->lt_invisible;
