+++ src/liboptical/sh_text.c
@@ -130,7 +130,7 @@ txt_transp_hook(const struct bu_structparse *sdp,
  * (there is no specific unload_datasource function).
  */
 HIDDEN int
-txt_load_datasource(struct txt_specific *texture, struct db_i *dbInstance, const unsigned long int size)
+txt_load_datasource(struct txt_specific *texture, struct db_i *dbInstance, const size_t size)
 {
     struct directory *dirEntry;
 
@@ -179,12 +179,12 @@ txt_load_datasource(struct txt_specific *texture, struct db_i *dbInstance, const
 
 	    /* check size of object */
 	    if (texture->tx_binunifp->count < size) {
-		bu_log("\nWARNING: %s needs %zu bytes, binary object only has %zu\n", bu_vls_addr(&texture->tx_name), size, texture->tx_binunifp->count);
+		bu_log("\nWARNING: %s needs %zu bytes, binary object only has %zu\n", bu_vls_addr(&texture->tx_name), (size_t)size, (size_t)texture->tx_binunifp->count);
 	    } else if (texture->tx_binunifp->count > size) {
 		bu_log("\nWARNING: Binary object is larger than specified texture size\n"
 		       "\tBinary Object: %zu pixels\n\tSpecified Texture Size: %zu pixels\n"
 		       "...continuing to load using image subsection...",
-		       texture->tx_binunifp->count, size);
+		       (size_t)texture->tx_binunifp->count, (size_t)size);
 	    }
 	}
     }
@@ -200,12 +200,12 @@ txt_load_datasource(struct txt_specific *texture, struct db_i *dbInstance, const
 	    return -1;				/* FAIL */
 
 	if (texture->tx_mp->buflen < size) {
-	    bu_log("\nWARNING: %s needs %zu bytes, file only has %zu\n", bu_vls_addr(&texture->tx_name), size, texture->tx_mp->buflen);
+	    bu_log("\nWARNING: %s needs %zu bytes, file only has %zu\n", bu_vls_addr(&texture->tx_name), (size_t)size, texture->tx_mp->buflen);
 	} else if (texture->tx_mp->buflen > size) {
 	    bu_log("\nWARNING: Texture file size is larger than specified texture size\n"
 		   "\tInput File: %zu pixels\n\tSpecified Texture Size: %zu pixels\n"
 		   "...continuing to load using image subsection...",
-		   texture->tx_mp->buflen, size);
+		   texture->tx_mp->buflen, (size_t)size);
 	}
 
     }
